Added calendar field parent attribute on form-open, middle-click navigation
Identical to list fields, calendar fields (incl. Gantt) can now use the record ID of the parent form to pre-fill an attribute when opening a new form.

Used when calendar field is placed as sub field and needs to create child-entities, which reference back to the parent.

Also enabled middle-click navigation for calendar day events and button fields.