Fixed relation filter policies nuking optional joins
Policy filters apply to all relations when accessed. The initial implementation worked well when working with a relation directly but had the unintended consequence of filtering out records when left-joining to a NULL relationship.

Now policy filters are applied on the join itself, still allowing for left-joins to work but correctly filtering out inaccessible records.