Fixed issue in data set API, updated comments
Issue occurs when a relationship attribute value is set during data set operation, but the relationship is joined as well. If a new record is being created, the relationship attribute value would be set twice (once as attribute value and once because of the joined relation).

This can happen if a writable relationship input field is added to a data form as well as attributes from the same, joined relation.