Reworked error message handling backend -> frontend
In most cases, errors should not occur on the backend as the frontend blocks invalid actions. But some actions cannot sensibly be caught - like foreign key violations on the database or parsing errors during CSV handling.

In these cases errors occur and must be communicate with, and handled by, the frontend. To display translated error messages with proper context and helpful information, some data only available to the backend must be passed to the frontend.

Before this commit, we recognized these "expected" backend errors via regexp and "just let them through", to be parsed again with regexp on the frontend for final transformation and display.

This commit introduces error codes. Messages are converted into codes on the backend with optional argument values for context. Error codes are then transformed and translated by the frontend.