Nested popup forms
Not really planned, but turns out we can easily enable nested pop-up forms. Nested pop-up forms can be useful when dealing with multiple levels of child records (or when checking other, related records).

We still need to check, whether we want to keep them and if so, need to refine them further.