Fix for collection retrieval error locking out admin users
If collections fail to load during login, the login attempt is aborted because something went wrong. This is by design. User should not access the system if it cannot successfully load as it would not work as expected.

However, because admins are also users, they could be locked out of the system, if they encounter an collection retrieval error.

Implemented solution: Regular users are still locked out of the system, but admins can still log-in (with prominent error message) to fix the collection issue.