Remove non-decryptable rows from results
Before this commit, when decrypting multiple rows, the entire result set was discarded if any value could not be decrypted.

Decryption can fail for many reasons - we should not throw away records that can be decrypted successfully. For now we remove the affected rows and log to the console for troubleshooting - we might want to implement some error message for this case (successful decryption of partial result set).