Frontend function: Form message, cleanup
New frontend function to show custom messages to users with a definable display timeout.

Also removed duplicate, shorter form messages that were used for mobile devices because of the limited screen space; with the previous commit we moved form messages to the title header, which gives us more space to work with.