List filter overwrite on first user change
Before, pre-configured list sort options were displayed and would need to be removed by the user when a new sorting is desired. This caused some confusion and was undesirable if multiple sort options were used as a default that needed to be removed.

Now pre-configured sort options are invisible and are immediately removed when a user chooses a different sort.