Fixed #33, workaround for edge-case
Thanks to @gizocz for the report.

To fix the issue, we needed to add the module name to field placeholders to guarantee uniqueness.

Also found an edge-case, in which the regular title for field references is not unique. If different modules with same-named relations have same-named relationship attributes with the same target, they share the same title.

Example (module.relation.attribute):
* assets.assignments.contact, (n:1) -> organization.contact
* absence.assignments.contact, (n:1) -> organization.contact
-> In form with "organization.contact" as base relation (0), they would both be shown as: "0) [1:n] assignments.contact", making unique identification difficult.

We should not just add the module name, as this would blow up all titles and is not necessary in 99% of cases. We implemented the following workaround: In addition to the title, we also add a mouse-over reference that includes the module name for unique identification.