Fixed: Menu still marked active for non-new record
When a menu entry opens a data form with a new record, it is marked as selected when this form is active. This is the default behavior and enables menu entries for "new record forms".

After a record has been created, the menu selection should however be removed as the record is not "new" anymore.