Attribute reference checks before deletion
Added reference checker for attributes, to resolve lingering references when attempting deletion.

It tries to collect all references that exist for the attribute, regardless on where (columns, fields, queries,  etc.) or how deeply nested they are (sub queries).

Referencing entities are then presented to the frontend for display and to enable navigation to the corresponding elements.