Removed reset of field settings on schema change
Because column aggregation settings on list fields were stored based on column order, thing could break if a list field changed. That is why we had the workaround to reset field settings on every schema change.

Now column aggregation is connected to column ID (instead of order). This means that schema changes at most cause invalid aggregation settings to not be used, instead of causing issues in list presentation.

Because of this change, we can now keep field settings alive even with schema changes.