Removed quick filter from field option cache
Regularly, user defined filters on list fields are cached as they adjust what is shown based on user choice.

We got feedback that this does not make sense for the quick filter, as its primary use case is to quickly filter down results in the moment. This commit removes the quick filter from the field option cache, effectively resetting it on every visit.

If a user wants to keep a filter permanently, they can use column or complex filters, which are still cached.