Fix: Rare error on record save with attached file
Issue occurred when trying to save a record with a newly uploaded file attachment, if the file was deleted by the scheduled file cleanup beforehand.

This was possible because the newly uploaded file did not have a reference yet (as record is not yet saved), so the cleanup job considered it eligible for removal.

The fix is to introduce a grace period for new files with no references. If there is only the first file version and this version is too young, the cleanup will skip it.